added benches for sign; benches and tests for keygen/sign refactored, added few more cases
