+++ src/gg_2018/mod.rs
@@ -16,4 +16,3 @@
 
 pub mod mta;
 pub mod party_i;
-pub mod test;
